Finish apis && endpoints stats
