--- stat-service/src/application/services/endpoints-stats/get-endpoints-calls-latency/index.ts
@@ -1,1 +1,0 @@
-export * from "./get-endpoints-calls-latency.service"
