--- stat-service/src/application/services/endpoints-stats/get-endpoints-error-calls/index.ts
@@ -1,1 +1,0 @@
-export * from "./get-endpoints-error-calls.service"
