--- stat-service/src/application/services/endpoints-stats/get-endpoints-success-calls/index.ts
@@ -1,1 +1,0 @@
-export * from "./get-endpoints-success-calls.service"
