+++ stat-service/src/presentation/api-stats/api-stats.controller.ts
@@ -3,7 +3,7 @@ import TryCatch from "../../infrastructure/api/try-catch.helper";
 import { get_api_stats_body_validator, get_api_stats_query_validator } from "../../contracts/api-stats/get_api_stats.request";
 import get_apis_stats_service from "../../application/services/api-stats/get_api_stats/get_api_stats.service";
 
-async function get_api_stats(req: Request, res: Response) {
+async function get_apis_stats(req: Request, res: Response) {
     return await TryCatch(res, async () => {
         const { api_ids } = get_api_stats_body_validator.parse(req.body)
         const { duration } = get_api_stats_query_validator.parse(req.query)
@@ -14,4 +14,4 @@ async function get_api_stats(req: Request, res: Response) {
     })
 }
 
-export default { get_api_stats }
+export default { get_apis_stats }
